+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new editor: `html_sanitize`
+
 ## [0.19.0] - 2022-05-11
 
 ### Added